refactor : IncomeCalculator 유틸 클래스로 변경
