+++ src/main/java/manager/BlackjackGameManager.java
@@ -6,7 +6,6 @@ import domain.participant.Dealer;
 import domain.participant.Name;
 import domain.participant.Player;
 import domain.result.Income;
-import domain.result.IncomeCalculator;
 import domain.result.Incomes;
 
 import java.util.LinkedHashMap;
@@ -36,7 +35,7 @@ public class BlackjackGameManager {
 
         printResults(dealer, playerBundle);
 
-        Incomes incomes = new Incomes(playerBundle, new IncomeCalculator());
+        Incomes incomes = new Incomes(playerBundle);
 
         printFinalIncomes(incomes.dealerIncome(dealer), incomes.playersIncomes(dealer));
     }
